Added new functionality

+++ src/main/java/org/fotum/app/commands/siege/settings/RemoveManagingRoles.java
@@ -1,18 +1,19 @@
-package org.fotum.app.commands.siege;
-
-import java.util.List;
-import java.util.concurrent.TimeUnit;
-
-import org.fotum.app.Constants;
-import org.fotum.app.features.siege.SiegeManager;
-import org.fotum.app.objects.ICommand;
+package org.fotum.app.commands.siege.settings;
 
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.TextChannel;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
+import org.fotum.app.Constants;
+import org.fotum.app.features.siege.GuildManager;
+import org.fotum.app.features.siege.GuildSettings;
+import org.fotum.app.utils.BotUtils;
+import org.fotum.app.objects.ICommand;
 
-public class RemoveManagingRole implements ICommand
+import java.util.ArrayList;
+import java.util.List;
+
+public class RemoveManagingRoles implements ICommand
 {
 
 	@Override
@@ -21,22 +22,28 @@ public class RemoveManagingRole implements ICommand
 		TextChannel channel = event.getChannel();
 		Member selfMember = event.getGuild().getSelfMember();
 		Member author = event.getMember();
-		
+
 		if (!author.hasPermission(Permission.MANAGE_SERVER))
 		{
-			this.sendMessageToChannel(channel, "You dont have permission to use this command");
+			BotUtils.sendMessageToChannel(channel, "You dont have permission to use this command");
 			return;
 		}
-		
-		Long guildId = event.getGuild().getIdLong();
+
+		long guildId = event.getGuild().getIdLong();
+		GuildSettings settings = GuildManager.getInstance().getGuildSettings(guildId);
+		if (settings == null)
+		{
+			BotUtils.sendMessageToChannel(channel, "No settings to remove for this guild");
+			return;
+		}
 		
 		if (selfMember.hasPermission(Permission.MESSAGE_MANAGE))
 		{
 			event.getMessage().delete().queue();
 		}
-		
-		SiegeManager.getInstance().removeManagingRole(guildId);
-		this.sendMessageToChannel(channel, "Managing role successfully removed");
+
+		settings.setManagingRoles(new ArrayList<Long>());
+		BotUtils.sendMessageToChannel(channel, "Managing roles successfully removed");
 	}
 
 	@Override
@@ -50,12 +57,5 @@ public class RemoveManagingRole implements ICommand
 	public String getInvoke()
 	{
 		return "remrole";
-	}
-
-	private void sendMessageToChannel(TextChannel channel, String msg)
-	{
-		channel.sendMessage(msg).queue(
-				(message) -> message.delete().queueAfter(5L, TimeUnit.SECONDS)
-		);
 	}
 }
